pyproject.toml
[project]
name = "langgraph-basics"
version = "0.1.0"
description = "LangGraph learning exercises"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "pydantic>=2.8",
    "python-dotenv>=1.0",
    "pytest>=9.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
